Optimization: Static emoji aliases
Previously the emoji dictionary got instantiated every time a string was transformed. Now the dictionary gets instantiated lazily the first it is needed and will be reused the next times.

As a bonus, this make the emoji aliases available to the outside via `[[NSString class] emojiAliases]` so this can be used i.e. for autocompletion.