add naïve reverse mapped dictionary of aliases 😇
The reversed dictionary is generated at runtime to avoid maintaining two distinct source lists. Duplicate emoji are trimmed through the use of NSSet and each is paired to a code by selecting the first result of `-[NSDictionary allKeysForObject:]`. It is important to note that the keys are not ordered and grouped sequentially as defined in `EMOJI_HASH` and in turn, will result in `:-1:` paired to 👎  and *not* `:+1:` paired to 👍  but instead `:thumbsup:`.